[vcpkg] Add --nuget-description option for `vcpkg export`
