+++ toolsrc/windows-bootstrap/vcpkg.vcxproj
@@ -205,6 +205,7 @@
     <ClInclude Include="..\include\vcpkg\base\unicode.h" />
     <ClInclude Include="..\include\vcpkg\base\util.h" />
     <ClInclude Include="..\include\vcpkg\base\view.h" />
+    <ClInclude Include="..\include\vcpkg\base\xmlserializer.h" />
     <ClInclude Include="..\include\vcpkg\base\zstringview.h" />
     <ClInclude Include="..\include\vcpkg\binarycaching.h" />
     <ClInclude Include="..\include\vcpkg\binaryparagraph.h" />
@@ -296,6 +297,7 @@
     <ClCompile Include="..\src\vcpkg\base\system.print.cpp" />
     <ClCompile Include="..\src\vcpkg\base\system.process.cpp" />
     <ClCompile Include="..\src\vcpkg\base\unicode.cpp" />
+    <ClCompile Include="..\src\vcpkg\base\xmlserializer.cpp" />
     <ClCompile Include="..\src\vcpkg\binarycaching.cpp" />
     <ClCompile Include="..\src\vcpkg\binaryparagraph.cpp" />
     <ClCompile Include="..\src\vcpkg\build.cpp" />
